api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
yfinance>=0.2.40
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.12.0
PyPortfolioOpt>=1.5.5
python-dotenv>=1.0.1
python-multipart>=0.0.9
openpyxl>=3.1.5
